R-randomForestSRC: Autospec creation for update from version 2.14.0 to version 3.0.0
CHANGES TO RELEASE 3.0.0
---------------------------------------------------------------------------------
Added the ability to get confidence intervals for out-of-bag error
rates using the subsample function.  Significant improvements in CPU
times for restoration of models and prediction with new data.  Fixed the
ability to select between in situ ensemble calculation and post-forest
ensemble calculation.  This can be useful when doing these
calculations in parallel causes significant contention for memory
locations that are in critical OpenMP regions.